asm memcpy can use shorter prefetch macro
Summary: It is just necessary to have a name which cannot collide. This can be done either by choosing a name in folly's namespace or by choosing any name that will not be otherwise used in the source. General mild preference is to the former but, in this case, style preference is for the latter.

Reviewed By: Orvid

Differential Revision: D34466323

fbshipit-source-id: 53d3fff137aa87643a055d31304e77451ab6dbac