docs: add GH button in support of Ukraine (#1728)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

Pull Request resolved: https://github.com/facebook/folly/pull/1728

Reviewed By: yfeldblum

Differential Revision: D34594831

Pulled By: dmitryvinn-fb

fbshipit-source-id: a8cda6445aad8ff230bc3c44c79d2f3e0ebbcbcc