folly: Remove deprecated synchronize_rcu function
Summary:
In folly RCU, a new `rcu_synchronize` API was added to replace the existing
`synchronize_rcu` API, in accordance with the new C++ Concurrency Technical
Standard 2 specification. This API has the exact same behavior as the
`synchronize_rcu` API, which itself simply wraps the new function. A series of
prior changes updated almost all callers in fbcode to use the new `rcu_synchronize`
API, with the exception of another proxygen revproxy usage which was added late
yesterday. This change removes `synchronize_rcu`, and updates the last caller in
proxygen to use `rcu_synchronize`.

Reviewed By: yfeldblum, paulmckrcu

Differential Revision: D34615411

fbshipit-source-id: 688f5a5d90c041f7e3112e4fe188bdde3d77d163