let coro::collectAllRange reserve its tasks vector
Summary:
It is tricky to get the size of an input range of unknown type. We must avoid iterating the range twice since iteration may be expensive.

But we do expect the input range to work with range-for, giving us options. Effectively, if the natural iterator type of the input range is a random-access iterator, then we use that as the signal that getting the size of the input range will be free.

Reviewed By: iahs

Differential Revision: D34540902

fbshipit-source-id: 3ac40ab19461e4222168d9db3e241d07b1404bc4