folly: Use concurrent updater threads in stress test
Summary:
The folly RCU Stress test currently does the following:

1. Creates an array of pointers to integers containing 0.
2. Creates an array of reader threads that iterate over the array and sum all
   of the values.
3. Creates a single updater that swaps pointers in the array, then invokes
   `rcu_retire` and updates + deletes the old swapped pointer.

The idea of course is to test that the pointer is not written to or freed by
the updater concurrently with any reader in a read region.

The stress test can be made a bit beefier by:

1. Having a pool of updater threads that are all concurrently updating the
   array, and freeing elements after epochs have elapsed.
2. Having the updaters use both `rcu_barrier` and `rcu_synchronize`, rather
   than just `rcu_barrier`.

This diff implements both of these ideas.

Reviewed By: paulmckrcu

Differential Revision: D34724916

fbshipit-source-id: f3e192fb6bef5a570e171b81cfc3064c347c7606