Use Huge Page Allocator in Zstd Context Pool
Summary:
When jemalloc is in use and the `JemallocHugePageAllocator` has been
initialized and there are huge pages available in its arena, zstd contexts
retrieved from the zstd compression context pool singleton will use huge
pages as the backing memory for their workspaces. When any of those
conditions don't hold, the allocator will fall back to just calling
`malloc()`.

In benchmarks, this produces a compression speedup of up to 15%.

Reviewed By: yfeldblum

Differential Revision: D29856163

fbshipit-source-id: 38e5f2e399c18524852a713589f0a8796c66aa1d