folly: Remove superfluous std::move's in rcu_reader_domain
Summary:
`folly::synchronization::rcu_reader_domain` implements a few types of
constructors, as well as a `swap` function. In some of these functions, there
is an `std::move` between two instances of the `rcu_reader_domain`. These
`std::move`'s are unnecessary. Two instances of `rcu_reader_domains` with the
same tag type must have the same pointer to an `rcu_domain`, due to the
requirement that an `rcu_domain` with some tag type must be globally unique.

Some of these `std::move`'s can therefore be replaced with more logical
semantics, such as exchanging for a `nullptr` for implementing a move
constructor, or verifying that two readers have the same domain for the
`swap` function.

Reviewed By: yfeldblum, paulmckrcu

Differential Revision: D34788674

fbshipit-source-id: e8c97101cb51932d4a64acdbdf37077a45329272