folly: Add more benchmark testcases for folly RCU
Summary:
folly RCU currently has two benchmarks:

1. A benchmark that tests creating an `rcu_reader` in a loop, and invoking
   `rcu_retire.`
2. A benchmark that just invokes `ThreadCachedInts::increment` in a loop.

There are a few other workloads we would like to measure. One is having nested
readers. That is, creating an `rcu_reader` inside of the scope of another
`rcu_reader`. There is functionally no difference between the two for the
current implementation of folly RCU, but we will be supporting only nested read
regions (as opposed to non-nested overlapping read regions) in a follow-on
change. Another is measuring `ThreadCachedInts` in different ways, such as
incrementing and decrementing in a tight loop, incrementing in a nested
context, and incrementing / decrementing in separate loops.

This change does two things:

1. Moves `//folly/test/ThreadCachedIntsBenchmark.cpp` to
   `//folly/synchronization/test/ThreadCachedIntsBench.cpp`, and renames the
   buck target to `thread_cached_ints_bench`. It also does the same for
   `//folly/synchronization/test/RcuBenchmark.cpp` and its target.

2. Adds these missing testcases to
   `folly/synchronization/test/RcuBenchmark.cpp` and
   `folly/test/ThreadCachedIntsBenchmark.cpp` respectively. They will be used
   to inform the upcoming new folly RCU implementation that provides fully
   nested read regions.
3. Rearranges the targets in `//folly/synchronization/test` to be alphabetic.

Reviewed By: yfeldblum

Differential Revision: D34755530

fbshipit-source-id: 18712e3925cc7bf4df489802337bec69794df798