folly::ext::test_find_resource
Summary:
An extension mechanism for folly tests to find their resources which can be hooked for build systems which need it.

Use it in places which need it, namely, `SingletonThreadLocal` and `Subprocess` tests.

Test resources are build artifacts which are associated with tests - in this case, helper libraries and binaries. The trouble is that build systems place them in arbitrary places which vary between build systems. So provide a mechanism which may be hooked.

Reviewed By: Gownta

Differential Revision: D34270239

fbshipit-source-id: 0e3c4dd72288083bbc7ec07ee12c48db74f8565c