Use Huge Page Allocator in Zstd Context Pool (Again)
Summary:
This is a resubmission of D29856163 (https://github.com/facebook/folly/commit/04c2275157a3e44c1d1fa1df75835fe4ec8a7b1e), which was reverted in D34848810 as a
result of test failures documented in T114204682. These failures seem to have
been caused by two issues.

It seems like in some modes, `JemallocHugePageAllocator::deallocate()` will
choke on addresses that were allocated via regular `malloc()` (while in other
situations it works??). So this resubmission adds a check before calling
`JEMHPA::deallocate()` to verify that the address is in the purview of the
JEMHPA.

Arguably this check should be in the `JemallocHugePageAllocator`'s impl of
`deallocate()` (and `reallocate()`), rather than be something the user has to
do. This is especially the case because JEMHPA will return normally
`malloc()`-ed addresses when the HPA arena has run out of space. I will likely
put up a follow-up that hoists this check into the JEMHPA implementations.

The second problem arose when the JEMHPA believed that jemalloc was in use at
build time, and therefore initialized `hugePagesSupported` to true. Since this
use of the JEMHPA does not call `JEMHPA::init()`, expecting the JEMHPA to fall
through to regular `malloc()` calls if not initialized, the JEMHPA never
validated that jemalloc was actually available. When this wasn't actually the
case at runtime, attempts to use non-available symbols, e.g. `mallocx()` would
then segfault.

This diff therefore adds a one-time check to the runtime `usingJEMalloc()` to
decide whether to pass the JEMHPA into zstd. Again, this is arguably a footgun
in the HPA itself, and the HPA should perhaps force a runtime check before it
tries to call jemalloc-specific symbols. Again I will think about whether this
warrants a follow-up diff.

Original Diff Summary:

When jemalloc is in use and the `JemallocHugePageAllocator` has been
initialized and there are huge pages available in its arena, zstd contexts
retrieved from the zstd compression context pool singleton will use huge
pages as the backing memory for their workspaces. When any of those
conditions don't hold, the allocator will fall back to just calling
`malloc()`.

In benchmarks, this produces a compression speedup of up to 15%.

Reviewed By: yfeldblum

Differential Revision: D34882676

fbshipit-source-id: 499e4d1f4f7679d586c9edf8311545e53ff7a309