Add support for signaling eventfds at runtime
Summary:
Add support for signaling eventfds at runtime.

Currently, we start by writing a large value of static constexpr uint64_t kNum = 2000000000; to the eventfd (it is a semaphore) so the eventfd is always io ready.

We want to benchmark the case where the eventds become io ready at runtime (the code path for the io_uring async read is different in this case).

Reviewed By: yfeldblum

Differential Revision: D34670464

fbshipit-source-id: 2aa124659e2d83c7f8c2ba2fce4db9754f4cdc4f