folly: Add ThreadCachedBench header for abstracting benchmarks
Summary:
There is currently a `//folly/test:thread_cached_ints_benchmark` benchmark that
is used to benchmark operations on the
`//folly/synchronization/detail:ThreadCachedInts.h` object. This object
provides APIs that allow folly RCU readers to enter/exit a read region by
incrementing and decrementing some specified epoch (potentially on different
threads), and allows RCU updaters to ensure that no readers are present in any
epoch on any thread.

This class will soon be replaced by an object that provides similar APIs to
both readers and updaters, though it ensures that read regions are fully nested
on a thread.

When comparing `ThreadCachedInts.h` to this new class, we'll want to be able to
run the exact same benchmarks on both objects. This diff therefore adds a new
`ThreadCachedBench.h` header file that benchmarks for `ThreadCachedInts`, and
the new object (likely to be called `ThreadCachedReaders`) can include.

As part of this change, the benchmarks are also moved into
`//folly/synchronization/test`, as `//folly/synchronization` is the only placed
they're used (and also where they're defined).

Reviewed By: yfeldblum

Differential Revision: D34787527

fbshipit-source-id: 67edb3f02430dfa2990bdb09e66cdda44e6c30da