folly: Add new ThreadCachedReaders class and benchmark
Summary:
There is currently a `folly::detail::ThreadCachedInts` class which provides
APIs that allow folly RCU readers to enter/exit a read region by incrementing
and decrementing some specified epoch (potentially on different threads), and
RCU updaters to ensure that no readers are present in any epoch on any thread.

This object has some advantages, and some drawbacks (as it relates to RCU):

Advantages:
- It is very fast, as the increment / decrement path are essentially just a
  single atomic increment (not including the decrement path's futex logic).
- The implementation is quite simple, as verifying that all readers have exited
  a read region for an epoch is simply a matter of summing the increments and
  decrements for either the even or odd epoch.

Disadvantages:
- It, by design, allows readers to enter and exit a read region on different
  threads. This is arguably a design flaw for an RCU API, as it makes debugging
  very difficult.
- It allows for non-nested, overlapping read regions. That is, two readers may
  overlap each other's read regions, with each belonging to a separate epoch.
  This has historically been a source of bugs in other RCU implementations, and
  is antithetical to the new C++ Concurrency TS2 standard being proposed in
  http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2021/n4895.pdf.

This diff introduces a new `folly::detail::ThreadCachedReaders` object which
addresses both of these disadvantages. Some aspects of the object's
implementation match that of `folly::detail::ThreadCachedInts`, such as using a
futex to signal to updaters that a reader has left a read region, and it can be
rechecked. The accounting for readers, however, is a bit different:

Rather than having two 2-element arrays of 64 bit atomics, one being an array
consituting readers entering/exiting the current even or odd epoch (i.e. an
array of increments), and one being the same for readers leaving (i.e. an array
of decrements), the `ThreadCachedReaders` object simply has a single 64-bit
thread local atomic that stores both the current epoch for the thread, as well
as the # of active, nested readers in the epoch.  When a reader enters a read
region, it checks if it is the first active reader, and if so, stores the
current epoch in the high order bits of the atomic. In all cases, the number is
incremented to reflect a new reader.  On the decrement path, the number is
simply decremented, and possibly the wait futex is triggered just as it is in
the `ThreadCachedInts` object.

An updater verifies that all readers have exited an epoch by iterating over all
threads' 64 bit atomic, and verifying that either the epoch is different than
the one being checked, or the number of readers is 0.

This addresses both of the disadvantages of `ThreadCachedInts`. A reader cannot
exit on a different thread, and read regions are fully nested. The
implementation for `ThreadCachedReaders` is also quite simple, though
performance is very slightly worse due to the increment / decrement logic being
slightly more complex than a single atomic increment as with
`ThreadCachedInts`.

Here are some numbers comparing the benchmarks:
`ThreadCachedReaders`:
```
$ buck2 run @//mode/opt //folly/synchronization/test:thread_cached_readers_bench
============================================================================
Benchmark                                                 time/iter  iters/s
============================================================================
IncrementLoop                                                1.79ns  558.00M
IncrementDecrementSameLoop                                   3.26ns  306.91M
IncrementDecrementSeparateLoop                               3.47ns  287.80M
----------------------------------------------------------------------------

$ buck run @//mode/opt //folly/synchronization/test:rcu_bench
============================================================================
Benchmark                                                 time/iter  iters/s
============================================================================
RcuReader                                                    2.78ns  359.95M
RcuReaderNested                                              2.90ns  345.33M
RcuRetire                                                   73.91ns   13.53M
----------------------------------------------------------------------------
```

`ThreadCachedInts`:
```
$ buck2 run @//mode/opt //folly/synchronization/test:thread_cached_ints_bench
============================================================================
Benchmark                                                 time/iter  iters/s
============================================================================
IncrementLoop                                                1.35ns  742.34M
IncrementDecrementSameLoop                                   1.35ns  742.84M
IncrementDecrementSeparateLoop                               3.22ns  310.29M
----------------------------------------------------------------------------

$ buck run @//mode/opt //folly/synchronization/test:rcu_bench
============================================================================
Benchmark                                                 time/iter  iters/s
============================================================================
RcuReader                                                    1.73ns  576.61M
RcuReaderNested                                              1.74ns  574.85M
RcuRetire                                                   77.93ns   12.83M
----------------------------------------------------------------------------
```

Reviewed By: yfeldblum

Differential Revision: D34788312

fbshipit-source-id: 541c2884fd7e179cf1f170acd4c025bcf4f8fa08