optimize 8-bit atomic fetch-bit-op
Summary:
Using 16-bit-wide `lock bt[src]` instructions and recalculating word addresses and bit offsets.

There are no 8-bit-wide `bt[src]` instructions, so we must get creative. While `lock (or|and|xor)` can work in cases where the exchanged value is unused, a CAS loop is required when it is used. And `lock bt[src]` may be faster than `lock (or|and|xor)`.

Reviewed By: Gownta

Differential Revision: D32240003

fbshipit-source-id: 083ae958c06869220ec334f0c7e7727755edb399