always_false
Summary:
This addresses a common shortcoming of how `static_assert` interacts with templated code. This pattern is used in Thrift (see `ServiceMetadata` for example). Currently that code uses `!sizeof(T)` which is makes use of the guarantee that `sizeof` is *almost* never 0 but that does not work with incomplete types (and the intent is less clear) .

[Here's a good blog post that explains in more detail](https://artificial-mind.net/blog/2020/10/03/always-false).

Reviewed By: iahs

Differential Revision: D34976834

fbshipit-source-id: 4ae97693d20b87879c8e503fc1c472cd6c523fc6