defer sync write behaviour to LogWriter
Summary: A simple write+flush is a reasonable default behaviour for synchronous writes, but for some custom async implementations of log writers, we may want to avoid locking the entire queue for a full flush(). This may require different implementations, different signatures, etc. So let's defer the decision of how to do synchronous writes to the writer, with write+flush as default behaviour.

Reviewed By: Gownta

Differential Revision: D35060544

fbshipit-source-id: cb20eaecdccee7b390e4270ddbcd4c0ea4b7a649