Add an option to prevent cancellation of .next() from closing the channel
Summary:
Currently, cancelling receiver.next() will close the channel for the given receiver. This has the advantage of matching the behavior of folly::coro::AsyncGenerator. However, it is both misleading to callers (who might expect only the wait for the next item to be cancelled), and unnecessarily limiting (preventing a caller from continuing to wait after cancelling).

This diff adds a new optional setting to avoid closing the channel upon cancellation of a call to .next(). Instead, only the call to .next() will be cancelled (as most users likely expect).

I will likely make this the only allowed behavior in a later diff (as a user can easily close the channel by destroying the receiver, something that will almost always be done automatically by cancelling the task doing the wait). However, I will want to check all current usages first, and I don't want that to block this stack (hence making this behavior optional for now).

Reviewed By: aary

Differential Revision: D35350405

fbshipit-source-id: 731480dd275fceca400f5b75c9d6e948e3688585