Save and restore rbx across cpuid on Windows
Summary:
We need to preserve `rbx` across the call to `cpuid` on Windows. This becomes obvious when compiling under ASan where presumably due to high register pressure, we end up in a situation where `rbx` is clobbered by `cpuid` but the compiler did not expect it to be clobbered.

This is fixed in clang upstream here:
https://github.com/llvm/llvm-project/issues/49477
https://reviews.llvm.org/D101338

However we use an older compiler (LLVM12.0.1) which does not have the patch yet.

Reviewed By: yfeldblum

Differential Revision: D35229538

fbshipit-source-id: 265835fb0e79e3a209dbce4fe82e8baa43e3d6ba