Drain error queue if ioReady looping
Summary:
It is possible that there are messages in the error queue yet `handleErrMessages()` returns without reading them because no error message callback is set and byte events are disabled. This could happen when a write is performed to an AsyncSocket with byte events enabled, triggers timestamps, and the fd is detached from the AsyncSocket and attached to a new AsyncSocket before the error messages generated for those timestamps arrive on the error queue. These `orphan` messages in the error queue will cause us to spin: `ioReady()` will be repeatedly invoked, but because we are not reading from the socket error queue, the state will never be cleared.

To prevent spinning under such circumstances, we drain the queue if the read or write handlers did not read or write anything on an invocation of `ioReady()`. Because these handlers can modify the event flags, we check these flags again before draining the error queue.

Reviewed By: bschlinker

Differential Revision: D34934824

fbshipit-source-id: 4c5518a2a090081ac4743613321b367c25cc334a