ChannelProcessor abstraction
Summary:
This object allows for memory-efficient processing of values many channels. A channel is added with a unique key and a callback. The callback will be called for every value pushed to the receiver.

A resumable channel can also be added. A resumable channel involves two callbacks. An initialization callback is called to get the receiver, and the update callback is called on every update (as for a normal channel). The update callback can throw a ReinitializeException at any time, which will trigger the initialize callback to re-run.

Values for a given channel are processed until one of the following occurs:
    1. The channel is closed
    2. The channel callback throws an OnClosedException
    3. The channel callback throws a folly::OperationCancelled exception.
    4. The channel is removed with a call to removeChannel.

If a channel is removed with removeChannel, processing will eventually stop for that channel. This will not necessarily happen immediately.

If a channel is added for an already existing key, the previous channel for that key will be removed and processing will eventually stop.

Processing for all channels will run on the user-provided executor. For any particular channel, all processing will happen sequentially. For any two distinct channels, processing may happen in parallel (subject to any constraints of the provided executor).

Reviewed By: nmagerko

Differential Revision: D35284916

fbshipit-source-id: 729d2a4d9082fc9bc1490f006a3db273c1c395a8