to_floating_point
Summary:
Like `to_integral`, but the other direction.

We want to enable integral-to-floating-point-conversion warnings. Explicit casts are okay, but conversions are warned. `to_floating_point` is a utility which clearly expresses the intention of conversion-to-floating-point. As v.s. an explicit cast to the target type, which does not clearly express the intention of the cast.

Differential Revision: D35857875

fbshipit-source-id: efa12917ec45ac4cead1a7de80acbb7956e10562