shave an instantiation per Function ctor instantiation
Summary:
The `Function` constructor accepting an arbitrary invocable tests that invocable for whether it may be considered empty. There is both a compile-time test and a runtime-test, wrapped up together in a function `isEmptyFunction`. But this function is instantiated for each instantiation of the `Function` constructor, and instantiations are expensive. The most common case is instantiating a `Function` with a lambda or other invocable which at compile-time cannot be considered empty. By lifting the compile-time test and applying it in a way which does not trigger a template instantiation, we can shave off some compile time.

```name=trunk
$ foundation/scripts/run-perf-compile --syntax folly/futures/test/FutureTest.cpp
     8,026,318,843      instructions:uP
```

```name=branch
$ foundation/scripts/run-perf-compile --syntax folly/futures/test/FutureTest.cpp
     8,000,507,359      instructions:uP
```

Reviewed By: ot, luciang

Differential Revision: D35624971

fbshipit-source-id: a1cf417fded06395ac607fdad8331030370a5641