folly: Remove template tag from ThreadCachedReaders.h
Summary:
`//folly/synchronization/detail/ThreadCachedReaders.h`, like `Rcu.h` and
`ThreadCachedLists.h`, is currently templatized on a tag that is propagated all
the way from users of the folly RCU library. The new TS2 standard described in
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2021/n4895.pdf does not
include a tag, so we'll want to remove tags from the public folly RCU API. This
change takes the first step towards doing that by removing it from
`folly::detail::ThreadCachedReaders`.

A tag will still required internally by `ThreadCachedReaders` and
`ThreadCachedLists`, as they use `folly::ThreadLocalPtr::accessAllThreads()`
which verifies that the tag on the `ThreadLocalPtr` is not `void`. To
accommodate this, we'll create a tag that is used by each in a shared header.

As part of this change, we also update `ThreadCachedReaders` to not use `static
thread_local` for storing per-thread state for RCU readers. The existing
mechanism used a `ThreadLocalPtr` that would store a lazily allocated
`EpochCount` pointer which was also cached in a `static thread_local
EpochCount*`. Now that the class is no longer templatized on a tag (which
creates new `static thread_local`'s per tag instance), we must query the
`ThreadLocalPtr` class directly, as it is now responsible for providing
per-instance TLS semantics.

Reviewed By: paulmckrcu

Differential Revision: D35056348

fbshipit-source-id: 18ec30d6d98f99b7ca46144ecbd8bbfc1a0b62ed