folly: Remove template tag from ThreadCachedLists.h
Summary:
`//folly/synchronization/detail/ThreadCachedLists.h`, like `Rcu.h`, is
currently templatized on a tag that is propagated all the way from users of the
folly RCU library. The new TS2 standard described in
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2021/n4895.pdf does not
include a tag, so we'll want to remove tags from the public folly RCU API. A
prior change removed the Tag template from
`folly::detail::ThreadCachedReaders`. This change similarly removes the Tag
template from `folly::detail::ThreadCachedLests`.

As part of this change, we also update `ThreadCachedLists` to not use `static
thread_local` for storing per-thread state for the lists. The existing
mechanism used a `ThreadLocalPtr` that would store a lazily allocated `TLHead`
pointer which was also cached in a `static thread_local TLHead*`. Now that the
class is no longer templatized on a tag (which creates new `static
thread_local`'s per tag instance), we must query the `ThreadLocalPtr` class
directly, as it is now responsible for providing per-instance TLS semantics.

With this, all of the internal data structures of `folly::rcu_domain` no
longer require a Tag, so we can remove it from the domain itself. This will be
done in a follow-on change, along with also removing the tag from all callers.

Reviewed By: paulmckrcu, Gownta

Differential Revision: D35056346

fbshipit-source-id: 26f4b2369930e0fd6af4128b97c0dfc791cce90d