folly: Remove Tag template from folly RCU
Summary:
The folly RCU library is currently templatized by a Tag. This was originally
done so as to allow us to use `static thread_local` variables in
`folly::detail::ThreadCachedInts` and `folly::detail::ThreadCachedLists`. While
it is very slightly more performant to use `static thread_local`'s, it is
functionally unnecessary with the use of `folly::ThreadLocalPtr`, which
provides per-instance thread local storage, and requires a branch and two moves
to `get()` the TLS pointer for the given instance. Benchmark numbers indicate
that performance is nearly identical between the two:

With Tag template:

Without Tag template (this change):

A series of prior changes removed the Tag template from `ThreadCachedReaders`
and `ThreadCachedLists`, leaving only the Tag template on the `rcu_domain`
itself, along with updating any callers. This change removes this final Tag
template, and updates all callers accordingly.

Reviewed By: Gownta

Differential Revision: D35056343

fbshipit-source-id: c010fa5cb14cee283009e7ffa290931bc389fde2