Optimize stringAppendfImpl, removing temporary buffers
The existing logic is, write to a 128 byte variable, if that's too small
then repeat that write to a larger heap variable. In both cases you must
then copy the results to the actual std::string. This is a minimum of two
writes and a maximum of 3 writes.

The new logic is similar except we write directly to the std::string for
a minimum of one write and a maximum of two writes. There is one less
copy and no duplicate strings in memory.

It's faster on my benchmarks, but I'm sure folly has their own
benchmarks also.