add missing sanitizer suppression in MicroLock
Summary:
`MicroLock` has 1 byte of state but operates on the containing aligned 4-byte word, principally because the linux `futex` syscall which backs it operates on the containing aligned 4-byte word.

This means that every operation which loads or stores this word must be done in a context with sanitizers suppressed.

Differential Revision: D36738948

fbshipit-source-id: 755d9557068d2efa9a74a9dd767a9fa5b3662d0c