Do not reserve() in sorted_vector_types bulk_insert
Summary:
`insert()` should never `reserve()` the final size, or it can trigger quadratic behavior in a sequence of constant-size range insertions. A good rule of thumb is that `insert(&item, &item + 1)` should behave the same as `insert(item)`.

Instead, delegate to the underlying container's `insert()`, which does the right thing in all reasonable implementations (including libstdc++'s `std::vector`): as soon as the iterator is at least `ForwardIterator`, it calls `std::distance` on it and reserves *at least* the final size, but accounting for the growth schedule.

Reviewed By: yfeldblum

Differential Revision: D36877827

fbshipit-source-id: de75d04ff0f0d655e82ce6faa9e2ea159f2f5a8b