Clamp small_vector capacity at maximum size.
Summary:
We already clamped the newSize at the maximum size of the vector,
however using goodMallocSize() could cause us to use a slightly larger capacity
than requested. If this exceeded maximum size of the vector (the maximum size
we can represent in the SizeType), we can end up with memory corruption if the
size is exceeded, rather than throwing an exception.

Reviewed By: mjoras

Differential Revision: D36818796

fbshipit-source-id: 4176b763b54bcbfffdd497d3218b777473c9854b