use iterator_has_known_distance_v in heap_vector_container
Summary:
Instead of replicating it with `distance_if_multipass`.

This changes semantics. Previously, if the iterator type was forward-iterator, the code would attempt `std::distance`. Now, this only happens if the iterator type is random-access-iterator. The motivation is that library code should not assume it is inexpensive to iterate twice - an example case may be a very expensive pipeline from range-v3.

Reviewed By: Gownta, ot

Differential Revision: D36890611

fbshipit-source-id: d3e8e92945cbd1fc80f0994c5b2a4edd951e1e2b