Python asyncio executor: keep asyncio loop reference until all KeepAlive tokens are released
Summary:
It's possible the asyncio event loop is destroyed before other C++ objects holding a keepalive token of the AsyncioExecutor. When the loop is deleted, the weakref dict `loop_to_q` will start deleting NotificationQueueAsyncioExecutor and call the destructor of the wrapped C++ AsyncioExecutor. If other C++ resources are still holding a KeepAlive token, the destructor will be stuck in the while loop calling drive().

In this change, we bump the refcnt of the loop every time a new KeepAlive token is acquired and then decref when it is released, so we may always make sure the loop is alive until all the keepalive tokens are released.

Reviewed By: junzh0u

Differential Revision: D33783486

fbshipit-source-id: ea7be9bde8233487af4789f7135d51f15a637a4f