let exception_wrapper::type() return a pointer
Summary:
Instead of returning a reference and using a sentinel referent to signal nullity.

When called on an empty `exception_wrapper` instance, or when called on an instance holding an object of type unrelated to `std::exception` on platforms where fast access is not supported, returns `nullptr`.

Phrased a different way: when `exception_wrapper::type()` returns `nullptr`, it does not necessarily mean that the `exception_wrapper` instance is empty! It may instead mean that the instance holds an object but cannot see the object's type.

Differential Revision: D37011994

fbshipit-source-id: 462373be849725ddefb2b3404b1a533afbe0db4f