Fix the cu index iteration.
Summary:
```
The index section header contains four unsigned 32-bit values (using the byte order of the application binary):

The version number of the format of this index (currently 2)
L, the number of columns in the table of section offsets
N, the number of compilation units or type units in the index
M, the number of slots in the hash table
(We assume that N and M will not exceed 232.)

The size of the hash table, M, must be 2k such that 2k > 3 * N / 2.

The hash table begins at offset 16 in the section, and consists of an array of M 64-bit slots. Each slot contains a 64-bit signature (using the byte order of the application binary).
```

Fix to use M instead of N.

also added boundary check for ranges section.

Reviewed By: yfeldblum

Differential Revision: D37132761

fbshipit-source-id: 0f9e955d00d3c6f94534b589ff7870df110ca360