Introduce flag to control number of threads in GlobalIOExecutor
Summary:
Currently `folly::getGlobalIOExecutor` returns IOThreadPoolExecutor which has same number of threads as number of cores, it's computed with `folly::hardware_concurrency()` and there is no easy way of overriding this parameter.

Introducing flag `io_threadpoolexecutor_thread_count` which is by default initialised to `0` in which case we still use number of cores as thread count but if it's set to some **non-zero positive value** we use it as a number of threads argument for IOThreadPoolExecutor.

We also added similar flags for `folly::getCPUExecutor` (CPUThreadPoolExecutor)

Reviewed By: yfeldblum

Differential Revision: D37087540

fbshipit-source-id: 18e04c4e55d13f6153fd546430e16bf186c49f5c