Remove fibers::ExecutionObserver::runnable
Summary:
`runnable` callback was added to fibers ExecutionObserver in D2081306
to measure fiber execution latency. Since then, not a single meaningful user of
this API exists.

On the other hand, the way `runnable` callback is implemented, opens
possibility for race conditions (Fiber::fiberManager_ is being accessed from
multiple threads without synchronization).

There is no easy way to preserve semantics of this callback while guaranteeing
thread safety and not hurting performance, so as the alternative let's remove
the feature altogether.

Reviewed By: yfeldblum

Differential Revision: D37804391

fbshipit-source-id: 70a191bcd2628fa484c2243af21dc237a8746701