Add support for positional args to NestedCommandLineApp
Summary:
Adds the ability to add positional args to subcommands on NestedCommandLineApp

https://www.boost.org/doc/libs/1_56_0/doc/html/program_options/overview.html#idp344486864

This allows for positional arguments to be mapped to a name in the `variables_map` instead of having to parse them from the `args` vector. E.g.

```
  po::positional_options_description positionalArgs;
  positionalArgs.add("foo", /*max_count=*/1);
  app.addCommand("subcommand", "[args...]", "Do something", "Help text", func, positionalArgs)
    .add_options()
    ("foo", boost::program_options::value<int32_t>()->required(), "foo positional arg")
```
would allow for the foo arg to be specified via either keyword or positionally and would get mapped to "foo" within the `variables_map` object returned
```
cli subcommand 42
cli subcommand --foo 42
```

Behavior for apps that don't specify command line positional args remains unchanged

Reviewed By: yfeldblum, lnyng

Differential Revision: D37788335

fbshipit-source-id: 91644927aed6562617e4632fdad46eb35d70435c